Initial Commit
This project stems from the need to highlight some of the risks misconfigurations in the cloud can represent to organisations and allow to proactively pen-test your AWS Cloud infrastructure.

This project relies on the AWS CLI to be present on a linux device where ideally path settings for AWS CLI have already been configured.